requirements.txt
django
hubspot-api-client
mysql-connector-python
python-decouple
dotenv
djangorestframework
